feat: save last uid

+++ Core/src/main/java/org/by1337/bauction/util/id/UniqueIdGenerator.java
@@ -1,23 +1,23 @@
 package org.by1337.bauction.util.id;
 
 public class UniqueIdGenerator {
-    private int pos;
+    private long pos;
 
-    public UniqueIdGenerator(int pos) {
+    public UniqueIdGenerator(long pos) {
         this.pos = pos;
     }
 
     public UniqueIdGenerator() {
     }
 
-    public synchronized int nextId(){
+    public synchronized long nextId(){
         return pos++;
     }
     public synchronized void setPos(int pos) {
         this.pos = pos;
     }
 
-    public synchronized int getPos() {
+    public synchronized long getPos() {
         return pos;
     }
 }
